Changed the normal samples, add the fast ones and implementation of the knn method

+++ headers/DTW.h
@@ -18,6 +18,6 @@ float average(float *array, int begin, int end);
 void ewma(float *array, int size, float *result);
 float dtwDistance(float *ax, float *ay, float *az, int sa, float *bx, float *by, float *bz, int sb, int window);
 float LBKeogh(float *x, int sx, float *y, int sy, int r);
-int knn(int16_t inputX[], int16_t inputY[], int16_t inputZ[], int size);
+int knn(float inputX[], float inputY[], float inputZ[], int size);
 
 #endif /* DTW_H_ */
